bot/requirements.txt
python-dotenv==0.20.0
aiohttp==3.8.1
aiogram==2.21
pyaes==1.6.1
Pyrogram==2.0.59
PySocks==1.7.1
TgCrypto==1.2.4